requirements.txt
streamlit
langchain
pypdf
sentence-transformers
faiss-cpu
openai
sqlalchemy
pymysql
python-dotenv